requirements.txt
paddlepaddle-gpu
pyyaml
easydict
numpy
scipy
pillow
opencv-python
tqdm
scipy